Move ls processing into its own class
